refactor: change package name from ui to presentation
