+++ app/src/main/java/com/github/taasonei/randomfox/data/database/DatabaseFox.kt
@@ -3,7 +3,7 @@ package com.github.taasonei.randomfox.data.database
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.github.taasonei.randomfox.ui.model.FoxPhoto
+import com.github.taasonei.randomfox.presentation.model.FoxPhoto
 
 @Entity(tableName = "favourites")
 data class DatabaseFox(
