+++ app/src/main/java/com/github/taasonei/randomfox/presentation/model/Status.kt
@@ -1,4 +1,4 @@
-package com.github.taasonei.randomfox.ui.model
+package com.github.taasonei.randomfox.presentation.model
 
 sealed interface Status {
     object Success : Status
