+++ app/src/main/java/com/github/taasonei/randomfox/presentation/view/FavouritesListFragment.kt
@@ -1,4 +1,4 @@
-package com.github.taasonei.randomfox.ui.view
+package com.github.taasonei.randomfox.presentation.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,7 +10,7 @@ import androidx.lifecycle.coroutineScope
 import androidx.recyclerview.widget.RecyclerView.Adapter.StateRestorationPolicy
 import com.github.taasonei.randomfox.R
 import com.github.taasonei.randomfox.databinding.FragmentFavouriteListBinding
-import com.github.taasonei.randomfox.ui.viewmodels.FavouritesListViewModel
+import com.github.taasonei.randomfox.presentation.viewmodels.FavouritesListViewModel
 import com.google.android.material.snackbar.Snackbar
 import kotlinx.coroutines.launch
