+++ app/src/main/res/layout/fragment_favourite_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.view.FavouritesListFragment">
+    tools:context=".presentation.view.FavouritesListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:layout_margin="@dimen/margin_tiny"
