+++ app/src/main/res/layout/fragment_recent_image.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.view.RecentImageFragment">
+    tools:context=".presentation.view.RecentImageFragment">
 
     <include
         android:id="@+id/recent_image_card"
